% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmat.R
\name{nstepmat}
\alias{nstepmat}
\title{Compute n-th Step Transition Probability Matrix}
\usage{
nstepmat(x, k, n, byrow = TRUE)
}
\arguments{
\item{x}{Numeric vector of one-step transition probability matrix values.}

\item{k}{Number of states.}

\item{n}{Higher order (number of steps).}

\item{byrow}{Logical; fill matrix by row or by column.}
}
\value{
A list containing the one-step transition matrix and the n-step transition matrix.
}
\description{
Given a one-step transition probability matrix, compute higher-order transition
probabilities of a homogeneous Markov chain process.
}
