% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsim.R
\name{simmarkov}
\alias{simmarkov}
\title{Simulate a homogeneous Markov Chain Process}
\usage{
simmarkov(init, onestep, lengthmc, statespace)
}
\arguments{
\item{init}{initial unconditional probability distribution over the states.}

\item{onestep}{onestep transitional probability matrix}

\item{lengthmc}{length of the Markov chain process except state at t=0}

\item{statespace}{state space or state labels as of the order of init}
}
\value{
vector of statespace in simulated markov chain process
}
\description{
Simulate a homogeneous Markov Chain Process
}
