% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationary.R
\name{stationary_prob}
\alias{stationary_prob}
\title{Compute Stationary probability distribution}
\usage{
stationary_prob(onestep)
}
\arguments{
\item{onestep}{one step transition probability matrix}
}
\value{
returns stationary probability distribution
}
\description{
Compute Stationary probability distribution
}
