% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmat.R
\name{nstepmat}
\alias{nstepmat}
\title{Compute n-th Step Transition Probability Matrix}
\usage{
nstepmat(x, k, n, byrow = TRUE)
}
\arguments{
\item{n}{Higher order}

\item{byrow}{To fill the matrix row-wise or column-wise}

\item{onestep}{One step transition probability matrix}

\item{states}{Number of states}
}
\value{
a list containing one-step transition probability matrix and n-step transition
}
\description{
Given one-step transition probability matrix compute higher-order transition
probability of a homogeneous Markov chain process
}
