% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_markov_chain.R
\name{plot_markov_chain}
\alias{plot_markov_chain}
\title{Plot the transition diagram}
\usage{
plot_markov_chain(P, state_names = NULL, threshold = 0.01)
}
\arguments{
\item{P}{transition probability matrix}

\item{state_names}{Optional vector of state labels}

\item{threshold}{Minimum probability to display an edge}
}
\value{
Transition diagram
}
\description{
Plot the transition diagram
}
